+++ src/schemas/hackathons/index.js
@@ -1,0 +1,13 @@
+import enrollsInSchema from './enrollsInSchema.js';
+import hackathonFilterSchema from './hackathonFilterSchema.js';
+import hackathonSchema from './hackathonSchema.js';
+import technologiesSchema from './technologiesSchema.js';
+import themesSchema from './themesSchema.js';
+
+export {
+    enrollsInSchema,
+    hackathonFilterSchema,
+    hackathonSchema,
+    technologiesSchema,
+    themesSchema,
+};
