+++ src/schemas/users/changePassSchema.js
@@ -1,14 +1,14 @@
 import Joi from '@hapi/joi';
-import joiErrorMessages from './joiErrorMessages.js';
+import joiErrorMessages from '../joiErrorMessages.js';
 
 //////////////////////////////////////////////////////////////
 // Esquema de cambio de contraseña
 // Tiene la contraseña vieja, la nueva, y la repetición de la nueva
 /////////////////////////////////////////////////////////////
-const userSchema = Joi.object().keys({
+const changePassSchema = Joi.object().keys({
     oldPass: Joi.string().max(100).required().messages(joiErrorMessages),
     newPass: Joi.string().max(100).required().messages(joiErrorMessages),
     repeatNewPass: Joi.string().max(100).required().messages(joiErrorMessages),
 });
 
-export default userSchema;
+export default changePassSchema;
