+++ src/schemas/users/index.js
@@ -1,0 +1,6 @@
+import changePassSchema from './changePassSchema.js';
+import loginUserSchema from './updateUserSchema.js';
+import updateUserSchema from './updateUserSchema.js';
+import userSchema from './userSchema.js';
+
+export { changePassSchema, loginUserSchema, updateUserSchema, userSchema };
